# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API Collect&Verything",
        "version": "0.1"
    },
    "paths": {
        "/api/stores/{id}/closings": {
            "post": {
                "tags": [
                    "Stores"
                ],
                "summary": "Post a new store closing",
                "description": "Create a new store closing",
                "operationId": "addClosing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Store Id",
                        "required": true
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "From",
                        "required": true
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "To",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account created"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/stores/{id}/closings/{closing_id}": {
            "delete": {
                "tags": [
                    "Stores"
                ],
                "summary": "Delete a store closing",
                "description": "Soft delete a store closing",
                "operationId": "removeClosing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account id",
                        "required": true
                    },
                    {
                        "name": "closing_id",
                        "in": "path",
                        "description": "Closing id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Closing deleted"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}